refactor(ip): convert filters to components
