--- packages/manager/apps/dedicated/client/app/ip/filters/ip/ipFirewallRulePort/ipFirewallRulePort.js
@@ -1,7 +1,0 @@
-angular.module('Module.ip.filters').filter(
-  'ipFirewallRulePort',
-  () =>
-    function ipFirewallRulePortFilter(port) {
-      return (port && port.replace(/^eq /, '')) || '';
-    },
-);
