--- packages/manager/apps/dedicated/client/app/ip/filters/ip/ipPunycode/ipPunycode.js
@@ -1,7 +1,0 @@
-angular.module('Module.ip.filters').filter(
-  'ipPunycode',
-  () =>
-    function ipPunycodeFilter(reverse, decode) {
-      return reverse ? punycode[decode ? 'toUnicode' : 'toASCII'](reverse) : '';
-    },
-);
