+++ packages/manager/apps/dedicated/client/app/ip/filters/ipFirewallRulePort/index.js
@@ -1,0 +1,4 @@
+export default () =>
+  function ipFirewallRulePortFilter(port) {
+    return (port && port.replace(/^eq /, '')) || '';
+  };
