+++ packages/manager/apps/dedicated/client/app/ip/filters/ipPunycode/index.js
@@ -1,0 +1,4 @@
+export default () =>
+  function ipPunycodeFilter(reverse, decode) {
+    return reverse ? punycode[decode ? 'toUnicode' : 'toASCII'](reverse) : '';
+  };
